update int32 to decimal
basically i messed up

+++ Program.cs
@@ -6,11 +6,11 @@
         {
             Console.Title = "xpcalc";
             Console.Write("wanderer\n");
-            int w = Convert.ToInt32(Console.ReadLine());
+            decimal w = Convert.ToDecimal(Console.ReadLine());
             Console.Write("adventurer\n");
-            int a = Convert.ToInt32(Console.ReadLine());
+            decimal a = Convert.ToDecimal(Console.ReadLine());
             Console.Write("hero\n");
-            int h = Convert.ToInt32(Console.ReadLine());
+            decimal h = Convert.ToDecimal(Console.ReadLine());
             Console.WriteLine(w * 1000 + a * 5000 + h * 20000 + "xp");
             Console.Read();
         }
